+++ docs/Changelog.md
@@ -7,6 +7,7 @@
 * [#4440](https://github.com/TouK/nussknacker/pull/4440) Improvement: Better exception info handling at KafkaExceptionInfo.inputEvent,
   from now we will return here JSON with all context variables (including context parent tree)
 * [#4452](https://github.com/TouK/nussknacker/pull/4452) Ace editor bump 1.4.12 -> 1.23.0
+* [#4455](https://github.com/TouK/nussknacker/pull/4455) Improvement: Allow to run TestScenarioRunner in TestRuntime mode
 
 1.10.0 (29 Jun 2023)
 -------------------------